Bug fixes for header
Fixed a few bugs, the first is that when clicking on a book there was a warning for session_start so took that out.

The second was a functions call error which was due to including the header in author.php and edit_profile.php so echoing the header instead of including fixed it.